Preparativos DemoVideo

+++ frontend/src/components/ProductOptions.js
@@ -10,17 +10,19 @@ const ProductOptions = (props) => {
   const history = useHistory();
 
   async function deleteProduct() {
-    const res = await fetch(`http://localhost:3300/product/${id}`, {
-      method: "DELETE",
-      headers: {
-        "Content-Type": "application/json",
-        authorization: token,
-      },
-    });
-    const data = await res.json();
-    console.log(data);
+    if (window.confirm("¿Seguro que deseas borrar este producto?")) {
+      const res = await fetch(`http://localhost:3300/product/${id}`, {
+        method: "DELETE",
+        headers: {
+          "Content-Type": "application/json",
+          authorization: token,
+        },
+      });
+      const data = await res.json();
+      console.log(data);
 
-    history.push("/");
+      history.push("/");
+    }
   }
 
   function editProduct() {
